requirements.txt
torch
torchvision
numpy
pandas
tqdm
numpy
matplotlib
imageio
imagehash
pillow
scikit-learn
umap-learn
numba
rasterfairy

newspaper3k
mediacloud
requests
python-dotenv
